Google-Speech-to-Text-API/requirements.txt
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
google-api-core==1.14.2
google-auth==1.6.3
google-cloud-core==1.0.3
google-cloud-speech==1.2.0
google-cloud-storage==1.19.0
google-resumable-media==0.4.0
googleapis-common-protos==1.6.0
grpcio==1.23.0
idna==2.8
Naked==0.1.31
prompt-toolkit==1.0.14
protobuf==3.9.1
pyasn1==0.4.7
pyasn1-modules==0.2.6
Pygments==2.4.2
PyInquirer==1.0.3
pytz==2019.2
PyYAML==5.1.2
regex==2019.8.19
requests==2.22.0
rsa==4.0
six==1.12.0
urllib3==1.25.3
wcwidth==0.1.7